feat: result api and refactor

+++ backend/src/main/java/com/doodledoodle/backend/result/dto/response/UserResultResponseDto.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class UserResultResponseDto {
     private Long drawId;
-    private Long drawNo;
+    private Integer drawNo;
     private String imgUrl;
     private Float similarity;
 }
